feat(rag): give each retrieval step its own module
A technical buyer can open one file for ingest, chunking, hybrid search, or citations instead of reading the Streamlit app.
